requirements.txt
bcrypt==5.0.0
fastapi==0.123.5
pandas==2.3.2
psycopg2==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.3
streamlit==1.51.0
streamlit-option-menu==0.4.0
